arm64: dts: marvell: armada-3720-db: Add vmmc regulator for SD slot
By adding this regulator, a proper reset is done during boot. Without
this, the UHS failed to be detected after a warm reboot when the SD card
remained in the slot, then it fallback to an HS.

Note that the vmcc is supported by the xenon driver only with the
following fix: "mmc: sdhci-xenon: add set_power callback".

Signed-off-by: Gregory CLEMENT <gregory.clement@free-electrons.com>

[ upstream commit: c13604d9ddc24dd4e9b65cad2844b2b603391ac8 ]