ARM: dts: Add missing properties for omap4 control modules
On omap4, we are missing several ti,hwmods properties and IO
ranges for system control modules. These are needed by the SoC
interconnect code.

Note that this will only show up as a bug with "doesn't have
mpu register target base" boot errors when the legacy platform
data is removed.

In order to add these, we need to move omap4_pmx_wkup to be a
child of omap4_padconf_wkup.

On omap4 there are separate modules for control module and
control module pads. For control module core, we have this
already configured except for the missing ti,hwmods and reg
entries.

Cc: Mark Rutland <mark.rutland@arm.com>
Acked-by: Rob Herring <robh+dt@kernel.org>
Signed-off-by: Tony Lindgren <tony@atomide.com>

[ upstream commit: 1d6a332ae8b7d8eacb9fe4c0a09370550f0b3eee ]