ARM: dts: omap3-evm-37xx: Add common processor module support
This commit moves common OMAP3-EVM processor module device tree data
to a separate include file. This will allow for 'omap3-evm.dts' to use
device tree data that is unique to the OMAP3530 version of the
processor module, while making use of the work already done for the
'omap3-evm-37xx.dts'.

Signed-off-by: Derald D. Woods <woods.technical@gmail.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>

[ upstream commit: 294b9cb84195a9606a22a4f5d0e26b51da6c0ce1 ]