ARM: dts: da850: Add DSP node
The TI Davinci DA8xx family of SoCs have a single DSP subsystem
that is comprised of TI's standard TMS320C674x megamodule and
several blocks of internal memory (L1P, L1D and L2 RAMs). Add
the DT node for this DSP processor sub-system. The processor
does not have an MMU, and uses a chip-level signalling register
and shared memory for inter-processor communication with the
ARM core.

The node has been added in disabled state, and can be enabled
in the respective board dts file with an associated reserved
memory block.

Signed-off-by: Suman Anna <s-anna@ti.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>

[ upstream commit: 21e7daef7e5e6fc79e8d7d84651ccc9a07b22042 ]