dt-bindings: fec: Make the phy-reset-gpio polarity explicit
The GPIO polarity passed to phy-reset-gpio is ignored by the FEC
driver and it is assumed to be active low.

It can be active high only when the 'phy-reset-active-high' property
is present.

The current examples pass active high polarity and work fine, but
in order to improve the documentation make it explicit what the real
polarity is.

Signed-off-by: Fabio Estevam <fabio.estevam@nxp.com>
Acked-by: Fugang Duan <fugang.duan@nxp.com>
Signed-off-by: David S. Miller <davem@davemloft.net>

[ upstream commit: 24fd3190817e96b2a57531986bfb6e951df3ee96 ]