ARM: dts: sun8i: fix USB Ethernet of Orange Pi R1
Orange Pi R1 uses a Realtek RTL8152B USB Ethernet chip, which is easily
seen on the board but not show in the schematics. A regulator for the
power of the RTL8152B chip is hidden, which uses the same pin with the
Wi-Fi regulator on the original Orange Pi Zero.

Add this regulator back to the device tree, and bind it to USB1.

Tested-by: Hauke Mehrtens <hauke@hauke-m.de>
Signed-off-by: Icenowy Zheng <icenowy@aosc.io>
Signed-off-by: Chen-Yu Tsai <wens@csie.org>

[ upstream commit: 532c129e11788954a4fc74baf6b787aa2cff6a67 ]