ARM: dts: sunxi: Change sun7i-a20-olimex-som204-evb to not use cd-inverted
Commit 45e01f401a2a16 ("ARM: dts: sunxi: Switch MMC nodes away from
cd-inverted property") changed most of the sunxi boards away from using
the cd-inverted property in MMC nodes. However, the
sun7i-a20-olimex-som204-evb board which got merged concurrently with
that commit is now using cd-inverted. Switch it away from using
cd-inverted to be consistent with rest of the sunxi boards.

Signed-off-by: Tuomas Tynkkynen <tuomas@tuxera.com>
Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>

[ upstream commit: c9e12d861395c87bf65707b88deca9e2b5f00c25 ]