ARM: dts: am3517-evm: Split off SOM features from baseboard
The am3517-evm consists of an am3517 SOM-M2 and a baseboard.
As items are added that are unique to the baseboard, let's place
them in a file called am3517-som.dtsi.  These items will be
common for all variants of the SOM.

Anything unique to the baseboard will be placed into
am3517-evm.

Signed-off-by: Derald D. Woods <woods.technical@gmail.com>
Signed-off-by: Adam Ford <aford173@gmail.com>
Signed-off-by: Tony Lindgren <tony@atomide.com>

[ upstream commit: 2500ee634f50c345184777eab8f3bd9b6f52857a ]