ARM: dts: da850-evm: Enable LCD and backlight
When using the board files the LCD works, but not with the DT.
This adds enables the original da850-evm to work with the same
LCD in device tree mode.

The EVM has a gpio for backlight regulator and a PWM for dimming
the backlight.  The LCD and the vpif display pins are mutually
exclusive, so if using the LCD, do not load the vpif driver.

Signed-off-by: Adam Ford <aford173@gmail.com>
Signed-off-by: Sekhar Nori <nsekhar@ti.com>

[ upstream commit: 8ae024497b3d68165e692d6ad5ba149cd070dac2 ]