ARM: dts: imx7d-sdb: Restore pwm backlight support
This was removed in commit 5eaeaccdaedb ("ARM: dts: imx7d-sdb: Pass
'enable-gpios' and 'power-supply' properties") with a note claiming that
GPIO1_IO01 is connected to a pin labeled as "PWREN" on the connector.

Despite that label this pin does actually work as a PWM controlling
brightness. So restore pwm functionality.

Signed-off-by: Leonard Crestez <leonard.crestez@nxp.com>
Reviewed-by: Fabio Estevam <fabio.estevam@nxp.com>
Signed-off-by: Shawn Guo <shawnguo@kernel.org>

[ upstream commit: 4a23e6ee9f691cfbd6bfc99c2c5aa8d23d8c1662 ]