ARM: dts: exynos: Add missing PMIC interrupts on UniversalC210 board
PMIC node requires external interrupts for proper operation. Lack of them
was probably caused by a limited support for interrupts from more than one
parent, which was not possible to define when support for this board was
added. Fix this by adding 'interrupt-extended' property and properly
define both interrupt lines for PMIC node.

Signed-off-by: Marek Szyprowski <m.szyprowski@samsung.com>
Signed-off-by: Krzysztof Kozlowski <krzk@kernel.org>

[ upstream commit: c58ce8e66d892b4131672bd875fe4eb0ed44c6a1 ]