ARM: dts: meson8b: odroidc1: add the fixed voltage regulators
There are multiple fixed regulators on the Odroid-C1 board. Add them so
they can be used when we add the devices that need them (SAR ADC needs
the 1.8V IOREF, RTC needs VDD_RTC).
These are:
- P5V0 is the main 5V power input
- VCC3V3 / VDDIO_AO3V3 / VDD3V3: fixed regulator with 3.3V output which
  is supplied by P5V0
- IOREF_1V8 / VCC1V8 / VDD1V8: fixed regulator with 1.8V output which is
  supplied by P5V0
- VDD_RTC: fixed voltage regulator with 0.9V output which is supplied by
  VDDIO_AO3V3
- DDR_VDDC / DDR3_1V5: fixed voltage regulator with 1.5V output which is
  supplied by P5V0
- the existing TF_IO and RFLASH_VDD_EN regulators are supplied by
  VDDIO_AO3V3
- the existing VCCK regulator is supplied by P5V0

This does not add the missing VDDEE regulator (controlled by PWM_D)
because it's not clear yet how to configure the voltage of that
regulator.

Signed-off-by: Martin Blumenstingl <martin.blumenstingl@googlemail.com>
Signed-off-by: Kevin Hilman <khilman@baylibre.com>

[ upstream commit: 288cb5d1dbbec52a5a9ca1814fd05102dcfa87be ]