dt-bindings: i2c: rcar: Add r8a77470 support
Although the I2C IP found in the RZ/G1C is not exactly the same
as the one found in the R-Car Gen2 family or R-Car Gen3 family,
it can still be considered as compatible with R-Car Gen2 from
a software perpective.
This patch therefore documents the SoC specific compatible string,
and the compatibility with Gen2 fallback is retained.

Signed-off-by: Fabrizio Castro <fabrizio.castro@bp.renesas.com>
Reviewed-by: Biju Das <biju.das@bp.renesas.com>
Signed-off-by: Wolfram Sang <wsa@the-dreams.de>

[ upstream commit: f4089fcfdb3d31421bbe0b1f3c3d6c11cd0dcb4d ]