ARM: dts: sun8i-h3: Add dts for the Mapleboard MP130
The Mapleboard MP130 is a single board computer based on the Allwinner
H3 SoC, with all schematics freely available. The Lite version includes
1GB main memory and 8GB eMMC.

https://www.mapleboard.org/en (still mostly in Chinese even when English
is selected)

This DTS is based upon the DTS shipped with the board which uses
mapleboard,mp130- prefixes instead of the allwinner,sun8i variants.

v2: Fold in review comments from Maxime Ripard

Signed-off-by: Jonathan McDowell <noodles@earth.li>
Signed-off-by: Maxime Ripard <maxime.ripard@bootlin.com>

[ upstream commit: 4f16ca40de15cd1073d5e9540d8e810deeab7d87 ]