ARM: dts: Adds device tree file for McGill's IceBoard, based on TI AM3874
This is an ARM + FPGA instrumentation board used at telescopes in
Antarctica, Chile, and Canada:

        https://pole.uchicago.edu/
        https://arxiv.org/abs/1608.03025
        https://chime-experiment.ca/

With these commits and a suitable userspace, we can boot the board, load
a FPGA bitstream, and communicate with the RTL design. Most of the board's
telemetry sensors (temperatures, voltages) are functional but detailed
testing is to follow.

We are weaning ourselves off TI's "official" kernel for this SOC, which
has been stuck at 2.6.37 and is not really fit for use. To anyone at TI:
despite good silicon and some dedicated support enginers, your
open-source software strategy for these parts has not worked well.
Please get in touch with me if you'd like to have a constructive
discussion about ways to improve it.

Signed-off-by: Graeme Smecher <gsmecher@threespeedlogic.com>
[tony@atomide.com: dropped fpga@1 as linux,spidev is still undocumented]
Signed-off-by: Tony Lindgren <tony@atomide.com>

[ upstream commit: d031773169df2de868c1b6ff75679ddac5d4a7ec ]