dt-bindings: display: tegra: Support SOR crossbar configuration
The SOR has a crossbar that can map each lane of the SOR to each of the
SOR pads. The mapping is usually the same across designs for a specific
SoC generation, but every now and then there's a design that doesn't.

Allow the crossbar configuration to be specified in device tree to make
it possible to support these designs.

Signed-off-by: Thierry Reding <treding@nvidia.com>

[ upstream commit: 6c2b3881d0df85fed7e3f43dcc9cbc3e5124bc12 ]