dt-bindings: at24: add Renesas R1EX24016
Document the compatible value for the Renesas R1EX24128ASAS0A two-wire
serial interface EEPROM, so it can be used in DTS files without causing
checkpatch warnings.

This is a 2 KiB EEPROM.  The first 1 KiB can always be written, the
second 1 KiB cannot be written if the write-protect line is asserted.

Signed-off-by: Geert Uytterhoeven <geert+renesas@glider.be>
Signed-off-by: Bartosz Golaszewski <bgolaszewski@baylibre.com>

[ upstream commit: d333bad1f6f3e0df224db181c991fb8d65bbc2be ]